Forth Commit
In this commit I downloaded many photos which I had not downloaded before and completed the first webpage with the insertion of 4 responsive images and background. I had to take out all images from image folder as the responsive images were not working even when I write the path to the image.